Add a string view class with unit tests
- Remove the fixed capacity option of the string class
- Increase the small buffer size of the string class
- Add common variable definitions for string classes unit tests

+++ unit_tests/gtest_staticstring_reverseiterator.cpp
@@ -129,7 +129,7 @@ TEST_F(StaticStringReverseIteratorTest, SubtractIndexToIterator)
 TEST_F(StaticStringReverseIteratorTest, SubtractIndexToIteratorInPlace)
 {
 	printf("Accessing characters with a reverse iterator and a negative index\n");
-	for (unsigned int i = 1; i <= static_cast<int>(string_.length()); i++)
+	for (unsigned int i = 1; i <= string_.length(); i++)
 	{
 		nctl::StaticString<Capacity>::ConstReverseIterator it = string_.rEnd();
